refactor(test): extract DovecotHarness into tests/support/dovecot
Pure relocation: tests/support/dovecot/{mod.rs,harness.rs} own the
container lifecycle helpers that e2e.rs declared inline. Test bodies
and assertions are unchanged; e2e.rs imports the moved harness via
#[path = "support/dovecot/mod.rs"] mod dovecot; use dovecot::DovecotHarness;.

DovecotHarness gains pub fn fingerprint() and pub fn port() accessors;
callers that previously reached into struct fields directly now use
those accessors instead.

Each test binary imports only the sub-module(s) it needs rather than
support/mod.rs wholesale, avoiding spurious dead-code warnings under
clippy::allow-attributes = "deny".

Sets up shared infrastructure Phase 3 (#265) builds on.
